refactor: fix cards files names by removing duplicating postfix (#4376)
