+++ .github/labeler.yml
@@ -27,7 +27,7 @@ lang-card:
   - changed-files:
       - any-glob-to-any-file:
           - api/top-langs.js
-          - src/cards/top-languages-card.js
+          - src/cards/top-languages.js
           - src/fetchers/top-languages.js
           - tests/fetchTopLanguages.test.js
           - tests/renderTopLanguagesCard.test.js
@@ -37,7 +37,7 @@ repo-card:
   - changed-files:
       - any-glob-to-any-file:
           - api/pin.js
-          - src/cards/repo-card.js
+          - src/cards/repo.js
           - src/fetchers/repo.js
           - tests/fetchRepo.test.js
           - tests/renderRepoCard.test.js
@@ -47,7 +47,7 @@ stats-card:
   - changed-files:
       - any-glob-to-any-file:
           - api/index.js
-          - src/cards/stats-card.js
+          - src/cards/stats.js
           - src/fetchers/stats.js
           - tests/fetchStats.test.js
           - tests/renderStatsCard.test.js
@@ -57,7 +57,7 @@ wakatime-card:
   - changed-files:
       - any-glob-to-any-file:
           - api/wakatime.js
-          - src/cards/wakatime-card.js
+          - src/cards/wakatime.js
           - src/fetchers/wakatime.js
           - tests/fetchWakatime.test.js
           - tests/renderWakatimeCard.test.js
@@ -67,7 +67,7 @@ gist-card:
   - changed-files:
       - any-glob-to-any-file:
           - api/gist.js
-          - src/cards/gist-card.js
+          - src/cards/gist.js
           - src/fetchers/gist.js
           - tests/fetchGist.test.js
           - tests/renderGistCard.test.js
