+++ src/cards/index.js
@@ -1,4 +1,4 @@
-export { renderRepoCard } from "./repo-card.js";
-export { renderStatsCard } from "./stats-card.js";
-export { renderTopLanguages } from "./top-languages-card.js";
-export { renderWakatimeCard } from "./wakatime-card.js";
+export { renderRepoCard } from "./repo.js";
+export { renderStatsCard } from "./stats.js";
+export { renderTopLanguages } from "./top-languages.js";
+export { renderWakatimeCard } from "./wakatime.js";
