+++ tests/gist.test.js
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 import { expect, it, describe, afterEach } from "@jest/globals";
-import { renderGistCard } from "../src/cards/gist-card.js";
+import { renderGistCard } from "../src/cards/gist.js";
 import { CONSTANTS, renderError } from "../src/common/utils.js";
 import gist from "../api/gist.js";
