+++ tests/wakatime.test.js
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 import wakatime from "../api/wakatime.js";
-import { renderWakatimeCard } from "../src/cards/wakatime-card.js";
+import { renderWakatimeCard } from "../src/cards/wakatime.js";
 import { expect, it, describe, afterEach } from "@jest/globals";
 
 const wakaTimeData = {
